requirements.txt
# Testing requirements
siphash-cffi
scapy

# Compliation requirements
jsl
jsonschema
pyinstaller